basic main menu and messaging screen
