+++ apps/wave-frontend/src/routes/messaging_screen/[slug]/+page.ts
@@ -1,0 +1,5 @@
+export const load = ({ params }) => {
+    return {
+        slug: params.slug,
+    };
+};
