+++ packages/ui/types/types.ts
@@ -1,11 +1,12 @@
 export type Conversation = {
     name: string;
-    id: string;
+    id: number;
 };
 
 export type Message = {
-    content: string;
-    sender: string;
-    sent_time_stamp: string;
-    id: string;
+    // sender: string;
+    // sent_time_stamp: string;
+    id: number;
+    text: string;
 };
+
